requirements.txt
fastapi
uvicorn
python-dotenv
PyGithub
requests
google-generativeai
openai
